feat: Added enum and custom store for controlling page content
