+++ src/components/Dashboard.svelte
@@ -1,0 +1,60 @@
+<script lang="ts">
+  import Connect from '$components/Connect.svelte';
+  import Capacity from '$components/Capacity.svelte';
+  import Provider from '$components/Provider.svelte';
+  import KeySelection from '$components/KeySelection.svelte';
+  import ProviderActions from '$components/ProviderActions.svelte';
+  import ChainStatus from '$components/ChainStatus.svelte';
+  import type { ChainInfo } from '$lib/storeTypes';
+
+  import {
+    storeBlockNumber,
+    storeChainInfo,
+    storeConnected,
+    storeMsaInfo,
+    storeToken,
+    storeValidAccounts,
+    transactionSigningAddress,
+  } from '$lib/stores';
+
+  const onChangeTxnSigningAddress = (evt: Event) => {
+    let option = evt.target as HTMLOptionElement;
+    storeMsaInfo.set({ isProvider: false, msaId: 0, providerName: '' });
+    transactionSigningAddress.set(option.value);
+  };
+
+  let token = '';
+
+  storeToken.subscribe((val) => (token = val));
+  let blockNumber = 0n;
+  let epochNumber = 0n;
+  let connected = false;
+  let validAccounts = {};
+
+  // TODO: put all this in chainInfo and update how it's stored.
+  storeBlockNumber.subscribe((val) => (blockNumber = val));
+  storeConnected.subscribe((val) => (connected = val));
+  storeValidAccounts.subscribe((val) => (validAccounts = val));
+  storeChainInfo.subscribe((info: ChainInfo) => (epochNumber = info.epochNumber));
+</script>
+
+<ChainStatus {blockNumber} {connected} {token} {epochNumber} />
+<div class="flex justify-center">
+  <Provider />
+  <Capacity bind:token />
+</div>
+<div class="mt-8 text-white">
+  <form id="setupForm">
+    <Connect />
+    <div class:hidden={!connected} class="mt-8">
+      <KeySelection
+        component="TransactionSigningKey"
+        selectLabel="Choose a Wallet Address"
+        selectedOption={''}
+        onSelect={onChangeTxnSigningAddress}
+        {validAccounts}
+      />
+    </div>
+  </form>
+  <ProviderActions {validAccounts} />
+</div>
