+++ src/lib/stores.ts
@@ -14,4 +14,24 @@ export const storeCurrentAction = writable(ActionForms.NoForm);
 
 export const storeToken = writable('');
 
-export const storeChainInfo = writable({ connected: false, blockNumber: 0n, epochNumber: 0n, token: ''});
+export const storeChainInfo = writable({ connected: false, blockNumber: 0n, epochNumber: 0n, token: '' });
+
+export enum PageContent {
+    Dashboard = 'dashboard',
+    Login = 'login',
+    Register = 'register'
+}
+
+const createPageContentStore = () => {
+    const { subscribe, set, update } = writable(PageContent.Login);
+
+    return {
+        subscribe,
+        set,
+        login: () => set(PageContent.Login),
+        register: () => set(PageContent.Register),
+        dashboard: () => set(PageContent.Dashboard)
+    };
+}
+
+export const pageContent = createPageContentStore();
