+++ src/routes/+page.svelte
@@ -1,59 +1,18 @@
 <script lang="ts">
-  import {
-    storeBlockNumber,
-    storeChainInfo,
-    storeConnected,
-    storeMsaInfo,
-    storeToken,
-    storeValidAccounts,
-    transactionSigningAddress,
-  } from '$lib/stores';
+  import bottomleft from '$lib/assets/bottom-left-bars.png';
+  import Dashboard from '$components/Dashboard.svelte';
+  import RequestToBeProvider from '$components/RequestToBeProvider.svelte';
   import Connect from '$components/Connect.svelte';
-  import Capacity from '$components/Capacity.svelte';
-  import Provider from '$components/Provider.svelte';
-  import KeySelection from '$components/KeySelection.svelte';
-  import ProviderActions from '$components/ProviderActions.svelte';
-  import ChainStatus from '$components/ChainStatus.svelte';
+  import { pageContent, PageContent } from '$lib/stores';
 
-  import type { ChainInfo, MsaInfo } from '$lib/storeTypes';
-
-  let token = '';
-  let blockNumber = 0n;
-  let epochNumber = 0n;
-  let connected = false;
-  let validAccounts = {};
-
-  // TODO: put all this in chainInfo and update how it's stored.
-  storeBlockNumber.subscribe((val) => (blockNumber = val));
-  storeToken.subscribe((val) => (token = val));
-  storeConnected.subscribe((val) => (connected = val));
-  storeValidAccounts.subscribe((val) => (validAccounts = val));
-  storeChainInfo.subscribe((info: ChainInfo) => (epochNumber = info.epochNumber));
-
-  const onChangeTxnSigningAddress = (evt: Event) => {
-    let option = evt.target as HTMLOptionElement;
-    storeMsaInfo.set({ isProvider: false, msaId: 0, providerName: '' });
-    transactionSigningAddress.set(option.value);
-  };
+  $pageContent = PageContent.Dashboard;
 </script>
 
-<ChainStatus {blockNumber} {connected} {token} {epochNumber} />
-<div class="flex justify-center">
-  <Provider />
-  <Capacity bind:token />
-</div>
-<div class="mt-8 text-white">
-  <form id="setupForm">
-    <Connect />
-    <div class:hidden={!connected} class="mt-8">
-      <KeySelection
-        component="TransactionSigningKey"
-        selectLabel="Choose a Wallet Address"
-        selectedOption={''}
-        onSelect={onChangeTxnSigningAddress}
-        {validAccounts}
-      />
-    </div>
-  </form>
-  <ProviderActions {validAccounts} />
-</div>
+{#if $pageContent === PageContent.Dashboard}
+<Dashboard />
+{:else if $pageContent === PageContent.Login}
+<Connect />
+{:else if $pageContent === PageContent.Register}
+<RequestToBeProvider />
+{/if}
+<img alt="decoration-bottom-left" src={bottomleft} class="mt-20 fixed bottom-0 left-8 -z-40"/>
